Fix translations countries (#3137)
* Update SchemaLocator.php

Replaces the use of a relative path with an absolute path. For example, when the generation of models in the cache is called from the Thelia php console, it uses the path local/config/shema.xml. Whereas if the generation of models in the cache is called from an http request, it uses the path web/local/config/shema.xml.

* Update SchemaLocator.php

fix code style

* Updated Italian country translations

* Updated 1 french country translation

* Suggestion to add a language we use

---------

Co-authored-by: hugo buffard <hugo.buffard@api-and-you.com>